remove Alluvial.Distributors namespace from Alluvial main assembly
