+++ Alluvial/Alluvial.csproj
@@ -53,10 +53,10 @@
     <Compile Include="Cursor{TPosition}.cs" />
     <Compile Include="DistributedMultiStreamCatchup{TData,TUpstreamCursor,TDownstreamCursor,TPartition}.cs" />
     <Compile Include="DistributorPipeAsync{T}.cs" />
-    <Compile Include="Distributors\ILease.cs" />
-    <Compile Include="Distributors\Leasable{T}.cs" />
-    <Compile Include="Distributors\Lease.cs" />
-    <Compile Include="Distributors\Lease{T}.cs" />
+    <Compile Include="ILease.cs" />
+    <Compile Include="Leasable{T}.cs" />
+    <Compile Include="Lease.cs" />
+    <Compile Include="Lease{T}.cs" />
     <Compile Include="Grouping.cs" />
     <Compile Include="IDistributedCatchup.cs" />
     <Compile Include="InMemoryDistributor.cs" />
