+++ Alluvial/DistributorPipeAsync{T}.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Threading.Tasks;
-using Alluvial.Distributors;
 
 namespace Alluvial
 {
