+++ Alluvial/Lease.cs
@@ -3,7 +3,7 @@ using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Alluvial.Distributors
+namespace Alluvial
 {
     /// <summary>
     /// A time-bound exclusive lease.
