+++ Alluvial/MultiStreamCatchup{TData,TUpstreamCursor,TDownstreamCursor}.cs
@@ -3,7 +3,6 @@ using System.Collections.Concurrent;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
-using Alluvial.Distributors;
 
 namespace Alluvial
 {
